new reestimate mod with several files - mod.rs, tests.rs and cache.rs
