+++ phylo/src/tkf_model/reestimate/tests.rs
@@ -1,12 +1,14 @@
 use approx::assert_relative_eq;
 use assert_matches::assert_matches;
+use fixedbitset::FixedBitSet;
 
 use crate::alignment::{AncestralAlignment, Sequences, MASA};
 use crate::alphabets::Alphabet;
 use crate::phylo_info::PhyloInfo;
 use crate::random::FakeGenerator;
-use crate::tkf_model::tests::setup_test_phylo;
-use crate::tkf_model::{EdgeSeqsReestimator, TKF92IndelCostBuilder};
+use crate::tkf_model::{
+    mapping_from_node_seq, tests::setup_test_phylo, EdgeSeqsReestimator, TKF92IndelCostBuilder,
+};
 use crate::{record_wo_desc as record, tree, Error};
 
 #[test]
@@ -73,4 +75,29 @@ fn tkf_reestimation_fails_for_leaf() {
 
     let err = reestimator.reestimate(&leaf_idx);
     assert_matches!(err, Err(Error::EdgeSeqsReestimator(msg)) if msg.contains("leaf"));
+}
+
+#[test]
+fn tkf_mapping_from_node_seq() {
+    let mut node_seq = FixedBitSet::with_capacity(5);
+    node_seq.insert(0);
+    node_seq.insert(2);
+    node_seq.insert(4);
+    let block_lens = [2, 3, 1, 4, 1];
+    let seq_len: usize = block_lens.iter().sum();
+    let expected_mapping = [
+        Some(0),
+        Some(1), // first block finished
+        None,
+        None,
+        None,    // second block finished
+        Some(2), // third block finished
+        None,
+        None,
+        None,
+        None,    // fourth block finished
+        Some(3), // fifth block finished
+    ];
+    let mapping = mapping_from_node_seq(&node_seq, &block_lens, seq_len);
+    assert_eq!(mapping, expected_mapping);
 }
